feat: Register pet tutor when adding a new pet
This commit adds the functionality to register a pet tutor when adding a new pet. The `RegisterPetTutor` method is called in the `PetsController` after saving the pet to the database. It retrieves the current user and associates them with the newly added pet by creating a new `AppUserPet` object and saving it to the context.

The `RegisterPetTutor` method takes the user ID and pet ID as parameters and creates a new `AppUserPet` object with the corresponding IDs. It then adds the object to the context and saves the changes asynchronously.

This feature allows for better tracking and management of pet tutors in the application.